requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.23
Werkzeug==0.9.4
argparse==1.2.1
gdata==2.0.18
itsdangerous==0.24
python-google-spreadsheet==2.0.0
wsgiref==0.1.2
